Move grouper_delta_user() into grouper
