+++ requiam/grouper.py
@@ -5,10 +5,14 @@ import pandas as pd
 from requests.exceptions import HTTPError
 
 from .commons import figshare_stem, figshare_group
+from .delta import Delta
+from .grouper_query import GrouperQuery
 
 from .logger import log_stdout
 
 # Administrative groups
+from .manual_override import update_entries
+
 superadmins = figshare_group('GrouperSuperAdmins', '', production=True)
 admins = figshare_group('GrouperAdmins', '', production=True)
 managers = figshare_group('GrouperManagers', '', production=True)
@@ -356,3 +360,75 @@ def create_active_group(group, grouper_dict, group_description=None, log=None, a
 
     create_groups(group, 'group_active', group_description, ga_test,
                   log0=log, add=add)
+
+
+def grouper_delta_user(group, stem, netid, uaid, action, grouper_dict,
+                       delta_dict, mo=None, sync=False, log=None,
+                       production=True):
+    """
+    Purpose:
+      Construct a Delta object for addition/deletion based for a specified
+      user. This is designed primarily for the user_update script
+
+    :param group: str
+      The Grouper group to update
+    :param stem: str
+      The Grouper stem (e.g., 'portal', 'quota')
+    :param netid: str
+      The User NetID
+    :param uaid: str
+      The User UA ID
+    :param action: str
+      The action to perform. 'add' or 'remove'
+    :param grouper_dict: dict
+      Dictionary containing grouper settings
+    :param delta_dict:
+      Dictionary containing delta settings
+    :param mo: ManualOverride object
+      For implementing change to CSV files. Default: None
+    :param sync: bool
+      Indicate whether to sync. Default: False
+    :param log: LogClass object
+      For logging
+    :param production: Bool to use production stem. Otherwise a stage/test is used. Default: True
+
+    :return d: Delta object class
+    """
+
+    if isinstance(log, type(None)):
+        log = log_stdout()
+
+    grouper_query = figshare_group(group, stem, production=production)
+    gq = GrouperQuery(**grouper_dict, grouper_group=grouper_query)
+
+    member_set = gq.members
+    if not isinstance(netid, list):
+        netid = [netid]
+    if not isinstance(uaid, list):
+        uaid = [uaid]
+    member_set = update_entries(member_set, netid, uaid, action, log=log)
+
+    d = Delta(ldap_members=member_set,
+              grouper_query_instance=gq,
+              **delta_dict,
+              log=log)
+
+    log.info(f"ldap and grouper have {len(d.common)} members in common")
+    log.info(f"synchronization will drop {len(d.drops)} entries to Grouper {group} group")
+    log.info(f"synchronization will add {len(d.adds)} entries to Grouper {group} group")
+
+    if sync:
+        log.info('synchronizing ...')
+        d.synchronize()
+
+        # Update manual CSV file
+        if not isinstance(mo, type(None)):
+            if production:
+                mo.update_dataframe(netid, uaid, group, stem)
+            else:
+                log.info("Working with figtest stem. Not updating dataframe")
+    else:
+        log.info('dry run, not performing synchronization')
+        log.info('dry run, not updating dataframe')
+
+    return d
